fix option 2
The minimal file system was buggy. The error did not appear
if the pod was running on localhost.

The defined directory path "https://ex.fs.com/test" was normalized by isomorphic-git to
"https:/ex.fs.com/test" ("//" can be normalized to "/").
Therefore the solid-auth-client did not send the request with credentials. Because the fetch
which is ran before handed added the origin in front of the path. Resulting in a request on
"https://some.origin.com/https:/ex.fs.com/test", which of course evaluated to a 404 Not Found status.

This is just a quick fix, by initialising the fs without the protocol,
the normalization does not effect the url and we can just add the protocol if needed.